Feat: User 영상 Score 추출 로직 구현 완료 #13
- User 영상을 업로드 하여 Landmark 추출
- User 영상의 챌린지 키워드를 찾아 동일한 Original 영상을 가져옴
- 두 영상의 Sync를 동일하기 맞춤
- 두 영상의 Landmark를 기반으로 Cosine Similarity 계산
- 두 영상의 0.1초당 유사도 score list 및 평균 score 계산
- 해당 데이터 DB에 등록

+++ dance_dance/challenges/models.py
@@ -32,7 +32,7 @@ class UserVideo(models.Model):
     thumbnail_image_url = models.CharField(max_length=512)
     uploaded_at = models.DateTimeField()
 
-    score = models.IntegerField(null=True)
+    score = models.FloatField(null=True)
     score_list = models.TextField(null=True)
     is_rank = models.BooleanField(default=False)
     # likes = models.ManyToManyField(User, related_name="like_videos")
